Limiting the impact of @Data

+++ src/main/java/land/face/strife/data/CombatDetailsContainer.java
@@ -37,12 +37,12 @@ public class CombatDetailsContainer {
       return null;
     }
     float totalWeight = 0;
-    for (double amount : skillWeight.values()) {
+    for (float amount : skillWeight.values()) {
       totalWeight += amount;
     }
     float xpTotal = totalExp * EXP_PERCENTAGE;
     xpTotal += WEIGHT_PERCENTAGE * totalWeight;
-    xpTotal *= 0.9f + Math.random() * 0.2f;
+    xpTotal *= (float) (0.9f + Math.random() * 0.2f);
     Map<LifeSkillType, Float> rewards = new HashMap<>();
     for (LifeSkillType type : skillWeight.keySet()) {
       rewards.put(type, Math.max(1, xpTotal * (skillWeight.get(type) / totalWeight)));
